@@ -19,15 +19,15 @@ func TestCreateSecret(t *testing.T) {
 	config := "dGVzdA=="
 
 	secret := struct {
-		APIVersion string                 `yaml:"apiVersion"`
-		Kind       string                 `yaml:"kind"`
-		Metadata   map[string]interface{} `yaml:"metadata"`
-		Type       string                 `yaml:"type"`
-		Data       map[string]string      `yaml:"data"`
+		APIVersion string            `yaml:"apiVersion"`
+		Kind       string            `yaml:"kind"`
+		Metadata   map[string]any    `yaml:"metadata"`
+		Type       string            `yaml:"type"`
+		Data       map[string]string `yaml:"data"`
 	}{
 		APIVersion: "v1",
 		Kind:       "Secret",
-		Metadata: map[string]interface{}{
+		Metadata: map[string]any{
 			"name":      name,
 			"namespace": namespace,
 		},
